Fix standalone component conversion issues from PR review

+++ gedbrowserng-frontend/src/app/modules/person/person-family-spouse.component.ts
@@ -5,7 +5,6 @@ import { PersonService, UserService } from '../../services';
 import { HasFamily } from '../../interfaces/has-family';
 import { PersonGetter } from './person-getter';
 import { RefreshPerson } from '../../interfaces';
-import { NgIf } from '@angular/common';
 import { RouterLink } from '@angular/router';
 import { MatIconButton } from '@angular/material/button';
 import { MatTooltip } from '@angular/material/tooltip';
@@ -22,17 +21,23 @@ import { MatIcon } from '@angular/material/icon';
  */
 @Component({
     selector: 'app-person-family-spouse',
-    template: `<span *ngIf="person" class="parent"><a
-    [routerLink]="['/' + dataset + '/persons', person.string]"
-    class="name">{{ person.indexName }} {{ lifespanYearString() }} [{{ person.string }}]</a>
-</span>
-<span *ngIf="person" class="example-fill-remaining-space"></span>
-<span *ngIf="hasSignedIn()">
-  <button mat-icon-button matTooltip="Unlink spouse" color="warn" (click)="unlink()">
-    <mat-icon matListIcon>link_off</mat-icon></button>
-</span>`,
+        template: `@if (person) {
+    <span class="parent"><a
+            [routerLink]="['/' + dataset + '/persons', person.string]"
+            class="name">{{ person.indexName }} {{ lifespanYearString() }} [{{ person.string }}]</a>
+    </span>
+}
+@if (person) {
+    <span class="example-fill-remaining-space"></span>
+}
+@if (hasSignedIn()) {
+    <span>
+        <button mat-icon-button matTooltip="Unlink spouse" color="warn" (click)="unlink()">
+            <mat-icon matListIcon>link_off</mat-icon></button>
+    </span>
+}`,
     styles: [],
-    imports: [NgIf, RouterLink, MatIconButton, MatTooltip, MatIcon]
+    imports: [RouterLink, MatIconButton, MatTooltip, MatIcon]
 })
 export class PersonFamilySpouseComponent extends PersonGetter
     implements OnInit, OnChanges {
